#79 refactor: 컴포넌트명과 파일명 통일
