+++ src/pages/main-page/sections/MonthlyChartSection.jsx
@@ -1,4 +1,4 @@
-import MonthlyChartHeader from '../components/ChartHeader';
+import MonthlyChartHeader from '../components/MonthlyChartHeader';
 import MonthlyChartTabs from '../components/MonthlyChartTabs';
 
 const MonthlyChartSection = () => {
